Project structure refractor
