+++ include/attack.h
@@ -1,0 +1,9 @@
+#ifndef ATTACK_H
+#define ATTACK_H
+
+#include "common.h"
+
+void ceri(uint32_t *, uint32_t, char, int);
+void import(Pair **, int, const char *);
+
+#endif
