+++ include/test_attack.h
@@ -1,0 +1,9 @@
+#ifndef TEST_ATTACK_H
+#define TEST_ATTACK_H
+
+#include "attack.h"
+
+void test_ceri();
+void test_import();
+
+#endif
