+++ include/test_common.h
@@ -1,0 +1,13 @@
+#ifndef TEST_COMMON_H
+#define TEST_COMMON_H
+
+#include "common.h"
+
+void print_hash(unsigned char *);
+void test_hash();
+void test_hash_n();
+void test_reduction();
+void test_hash_reduction();
+void test_dict();
+
+#endif
