+++ include/test_precomp.h
@@ -1,12 +1,9 @@
 #ifndef TEST_H
 #define TEST_H
 
-#include "../include/precomp.h"
+#include "precomp.h"
 
-void print_hash(unsigned char *);
-void hash_n(uint32_t);
-void test_reduction();
-void test_hash_reduction();
+
 void test_initialize();
 void test_generate();
 void test_swap();
@@ -14,10 +11,7 @@ void test_sort();
 void test_gen_sort();
 void test_clean();
 void test_rice();
-void test_ceri();
 void test_export();
-void test_import();
 void test_precomp();
-void test_dict();
 
 #endif
