+++ src/precomp.c
@@ -97,11 +97,6 @@ void rice(uint32_t *end, uint32_t value, char k)
     *end = ((1 << (value >> k)) - 1) << (k + 1) | (value & ((1 << k) - 1));
 }
 
-void ceri(uint32_t *end, uint32_t value, char k, int nb_bits)
-{
-    *end = ((nb_bits - k - 1) << k) | (value & ((1 << k) - 1));
-}
-
 void export(Points *table, int table_size, const char *file_name)
 {
     FILE *file;
@@ -111,7 +106,7 @@ void export(Points *table, int table_size, const char *file_name)
         exit(ERROR_FOPEN);
     }
 
-    if ((fwrite(table, sizeof(Points), table_size, file)) != (size_t) table_size)
+    if ((fwrite(table, sizeof(Points), table_size, file)) != (size_t)table_size)
     {
         fprintf(stderr, "Writing file problem : %s\n", file_name);
         exit(ERROR_FWRITE);
@@ -124,30 +119,9 @@ void export(Points *table, int table_size, const char *file_name)
     }
 }
 
-void import(Points *table, int table_size, const char *file_name)
+void precomp()
 {
-    FILE *file;
-    if ((file = fopen(file_name, "rb")) == (FILE *)NULL)
-    {
-        fprintf(stderr, "Opening file problem : %s\n", file_name);
-        exit(ERROR_FOPEN);
-    }
-
-    if ((fread(table, sizeof(Points), table_size, file)) != (size_t) table_size)
-    {
-        fprintf(stderr, "Reading file problem : %s\n", file_name);
-        exit(ERROR_FWRITE);
-    }
-
-    if (fclose(file))
-    {
-        fprintf(stderr, "Closing file problem : %s", file_name);
-        exit(ERROR_FCLOSE);
-    }
-}
-
-void precomp(){
-    int size = (int) ceil(m0);
+    int size = (int)ceil(m0);
     Points *table, *perfect;
     if ((table = (Points *)calloc(size, sizeof(Points))) == NULL)
     {
@@ -181,8 +155,8 @@ void precomp(){
     export(table, size, "./table2.dat");
     time_t e = time(NULL);
     printf("Time to export %d\t: %lds\n", size, e - c);
-    printf("Hash operations :\n\texpected\t: %d\n\texperimental\t: %d\n", (int) ceil(m0) * t, nb_hash);
-    printf("Unique endpoints :\n\texpected\t: %d\n\texperimental\t: %d\n", (int) ceil(mt), size);
+    printf("Hash operations :\n\texpected\t: %d\n\texperimental\t: %d\n", (int)ceil(m0) * t, nb_hash);
+    printf("Unique endpoints :\n\texpected\t: %d\n\texperimental\t: %d\n", (int)ceil(mt), size);
     printf("Table :");
     for (Points *current = perfect, *last = perfect + 20; current < last; current++)
         printf("\n%u\t:\t%u", current->start, current->end);
